requirements.txt
beautifulsoup4==4.11.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
db-dtypes==1.0.5
dnspython==2.3.0
google==3.0.0
google-api-core==2.11.0
google-auth==2.16.0
google-cloud==0.34.0
google-cloud-bigquery==3.5.0
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.1
grpcio-status==1.51.1
idna==3.4
motor==3.1.1
numpy==1.24.2
packaging==23.0
pandas==1.5.3
proto-plus==1.22.2
protobuf==4.21.12
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==4.3.3
python-dateutil==2.8.2
python-decouple==3.7
pytz==2022.7.1
requests==2.28.2
rsa==4.9
six==1.16.0
soupsieve==2.4
tornado==6.2
urllib3==1.26.14
